Zepto: DB layer for order created

+++ spring-feb-2025/zepto-clone/db-api/src/main/java/com/db/zepto/db_api/model/AppOrder.java
@@ -20,7 +20,7 @@ public class AppOrder {
     AppUser customer;
     @ManyToOne
     AppUser deliveryPartner;
-    int totalAmount;
+    double totalAmount;
     @OneToMany
     List<Product> products;
 }
